refactor: move item status
