+++ app/src/features/items/ItemLink.tsx
@@ -5,7 +5,7 @@ import { useStore } from '../../data/store'
 import { ItemId, ItemType } from '../../data/types'
 import { getSerialFromId, renderDate } from '../../data/util'
 import { ItemIcon } from '.'
-import { isActive, ItemStatus } from '../../components/Status'
+import { isActive, ItemStatus } from './Status'
 
 export const PATHS: { [type in ItemType]: string } = {
   issue: 'issues',
