+++ app/src/features/items/editor/IssueEditor.tsx
@@ -3,7 +3,7 @@ import { IssueIcon } from '../ItemIcon'
 import { Controller, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { IssueId, IssueStatus, NewIssue } from '../../../data/types'
-import { ISSUE_STATUS_UI } from '../../../components/Status'
+import { ISSUE_STATUS_UI } from '../Status'
 import { selectAllIssues, selectAllTags, useStore } from '../../../data/store'
 import { useEditorStore } from './ItemEditor'
 import { useNavigate } from 'react-router-dom'
