+++ app/src/features/items/index.ts
@@ -5,3 +5,4 @@ export * from './useNavigateToItem'
 export * from './editor'
 export * from './ItemIcon'
 export * from './ItemLink'
+export * from './Status'
