+++ app/src/pages/ImprovementsPage.tsx
@@ -2,7 +2,7 @@
 import { ImprovementStatus } from '../data/types'
 import { selectAllImprovements, useStore } from '../data/store'
 import { PageHeader } from '../features/layout'
-import { POSSIBLE_STATUSES } from '../components/Status'
+import { POSSIBLE_STATUSES } from '../features/items'
 import { ItemsTablePage } from '../features/items'
 import { Aim42ItemDescription } from '../features/aim42'
