+++ app/src/pages/IssuePage.tsx
@@ -1,7 +1,7 @@
 import { Tag, Row, Descriptions, Button, Typography } from 'antd'
 import { useStore } from '../data/store'
 import { EditOutlined, StopOutlined } from '@ant-design/icons'
-import { ISSUE_STATUS_UI } from '../components/Status'
+import { ISSUE_STATUS_UI } from '../features/items'
 import { renderDate } from '../data/util'
 import { IssueIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
