+++ app/src/pages/search.tsx
@@ -2,7 +2,7 @@ import { Input } from 'antd'
 import { useEffect, useMemo, useState } from 'react'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { PageHeader } from '../features/layout'
-import { getAllPossibleStatuses } from '../components/Status'
+import { getAllPossibleStatuses } from '../features/items'
 import { selectAllItems, useStore } from '../data/store'
 import { Item } from '../data/types'
 import { exists } from '../data/util'
